+++ pyodata/config.py
@@ -1,41 +1,8 @@
-""" Contains definition of configuration class for PyOData and for ODATA versions. """
+""" Contains definition of configuration class for PyOData"""
 
-from abc import ABC, abstractmethod
-from typing import Type, List, Dict, Callable, TYPE_CHECKING
-
+from typing import Type, Dict
 from pyodata.policies import PolicyFatal, ParserError, ErrorPolicy
-
-# pylint: disable=cyclic-import
-if TYPE_CHECKING:
-    from pyodata.model.elements import Typ, Annotation  # noqa
-
-
-class ODATAVersion(ABC):
-    """ This is base class for different OData releases. In it we define what are supported types, elements and so on.
-        Furthermore, we specify how individual elements are parsed or represented by python objects.
-    """
-
-    def __init__(self):
-        raise RuntimeError('ODATAVersion and its children are intentionally stateless, '
-                           'therefore you can not create instance of them')
-
-    # Separate dictionary of all registered types (primitive, complex and collection variants) for each child
-    Types: Dict[str, 'Typ'] = dict()
-
-    @staticmethod
-    @abstractmethod
-    def primitive_types() -> List['Typ']:
-        """ Here we define which primitive types are supported and what is their python representation"""
-
-    @staticmethod
-    @abstractmethod
-    def build_functions() -> Dict[type, Callable]:
-        """ Here we define which elements are supported and what is their python representation"""
-
-    @staticmethod
-    @abstractmethod
-    def annotations() -> Dict['Annotation', Callable]:
-        """ Here we define which annotations are supported and what is their python representation"""
+import pyodata.version
 
 
 class Config:
@@ -46,7 +13,7 @@ class Config:
     """ This is configuration class for PyOData. All session dependent settings should be stored here. """
 
     def __init__(self,
-                 odata_version: Type[ODATAVersion],
+                 odata_version: Type[pyodata.version.ODATAVersion],
                  custom_error_policies=None,
                  default_error_policy=None,
                  xml_namespaces=None
@@ -108,7 +75,7 @@ class Config:
         self._namespaces = value
 
     @property
-    def odata_version(self) -> Type[ODATAVersion]:
+    def odata_version(self) -> Type[pyodata.version.ODATAVersion]:
         return self._odata_version
 
     @property